refactor: upd zombie error folder name
