+++ packages/client/src/ui/fileInput/index.tsx
@@ -1,7 +1,7 @@
 import { useId, useRef, useState } from "react";
 import cn from "classnames";
 
-import ZombieError from "../zombie_error";
+import ZombieError from "../zombieError";
 import style from "./style.module.scss";
 
 type TProps = {
